TTALE-184: Vaguely working prototype for .war files
