+++ src/main/java/org/jboss/tattletale/core/NestableArchive.java
@@ -37,7 +37,7 @@ public abstract class NestableArchive extends Archive
    static final long serialVersionUID = -9197985968607581451L;
 
    /** Sub-archives */
-   private List<Archive> subArchives;
+   protected List<Archive> subArchives;
 
    /**
     * Constructor
@@ -64,6 +64,35 @@ public abstract class NestableArchive extends Archive
             classDependencies, packageDependencies, blacklistedDependencies, location);
 
       this.subArchives = null;
+   }
+
+   /**
+    * Constructor
+    *
+    * @param type                    The type
+    * @param name                    The name
+    * @param version                 The version number
+    * @param manifest                The manifest
+    * @param sign                    The signing information
+    * @param requires                The requires
+    * @param provides                The provides
+    * @param classDependencies       The class dependencies
+    * @param packageDependencies     The package dependencies
+    * @param blacklistedDependencies The blacklisted dependencies
+    * @param location                The location
+    * @param subArchives             The sub-archives that are stored locally.
+    */
+   public NestableArchive(int type, String name, int version, List<String> manifest, List<String> sign,
+                          SortedSet<String> requires, SortedMap<String, Long> provides,
+                          SortedMap<String, SortedSet<String>> classDependencies,
+                          SortedMap<String, SortedSet<String>> packageDependencies,
+                          SortedMap<String, SortedSet<String>> blacklistedDependencies, Location location,
+                          List<Archive> subArchives)
+   {
+      super(type, name, version, manifest, sign, requires, provides,
+            classDependencies, packageDependencies, blacklistedDependencies, location);
+
+      this.subArchives = subArchives;
    }
 
    /**
